Move SafeActionSpace to experimental, drop support for deprecated location
