+++ universe/wrappers/__init__.py
@@ -2,7 +2,6 @@ import gym
 import universe.wrappers.experimental
 from universe import envs, spaces
 from universe.wrappers import gym_core_sync
-from universe.wrappers.action_space import SafeActionSpace, SoftmaxClickMouse
 from universe.wrappers.blocking_reset import BlockingReset
 from universe.wrappers.diagnostics import Diagnostics
 from universe.wrappers.gym_core import GymCoreAction, GymCoreObservation, CropAtari
